Feat: Make EnvironmentID Optional (#60)
* feat: make environmentId optional

* chore: changeset
