+++ .changeset/long-buses-allow.md
@@ -1,0 +1,8 @@
+---
+'@flatfile/embedded-utils': patch
+'@flatfile/javascript': patch
+'@flatfile/react': patch
+'@flatfile/vue': patch
+---
+
+Make environmentId optional
